gb_backend_ebs/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
gunicorn==20.1.0
inflection==0.5.1
jmespath==1.0.1
pip==23.0.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
wheel==0.38.4
